requirements/base.txt
# Base
pytz==2018.9
python-slugify==3.0.2
Pillow==6.0.0
psycopg2==2.8.1 --no-binary psycopg2

# Django
django==2.2

# Django REST Framework
djangorestframework==3.9.2

# JWT
pyjwt==1.7.1

# Django Filter
django-filter==2.1.0

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==19.1.0

# Static files
whitenoise==4.1.2

# Celery
redis==3.2.1
django-redis==4.10.0
celery==4.3.0
flower==0.9.3
